Created Different Packages for Commands
Also changed stop command button to trigger (button1)
