+++ src/org/usfirst/frc/team166/robot/OI.java
@@ -4,10 +4,10 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.buttons.JoystickButton;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-import org.usfirst.frc.team166.robot.commands.CancelDriveCommand;
-import org.usfirst.frc.team166.robot.commands.LiftWings;
-import org.usfirst.frc.team166.robot.commands.LowerWings;
-import org.usfirst.frc.team166.robot.commands.StrafeDirection;
+import org.usfirst.frc.team166.robot.commands.drive.CancelDriveCommand;
+import org.usfirst.frc.team166.robot.commands.drive.StrafeDirection;
+import org.usfirst.frc.team166.robot.commands.wings.LiftWings;
+import org.usfirst.frc.team166.robot.commands.wings.LowerWings;
 import org.usfirst.frc.team166.robot.subsystems.Drive;
 
 /**
@@ -23,11 +23,10 @@ public class OI {
 	public OI() {
 
 		JoystickButton button1 = new JoystickButton(driveJoystick, 1);
-		JoystickButton button2 = new JoystickButton(driveJoystick, 2);
 		JoystickButton button3 = new JoystickButton(driveJoystick, 3);
 		JoystickButton button4 = new JoystickButton(driveJoystick, 4);
 
-		button2.whenPressed(new CancelDriveCommand());
+		button1.whenPressed(new CancelDriveCommand());
 		button3.whileHeld(new StrafeDirection(Drive.StrafeDirection.Left));
 		button4.whileHeld(new StrafeDirection(Drive.StrafeDirection.Right));
