+++ src/org/usfirst/frc/team166/robot/commands/autonomous/Autonomous.java
@@ -1,6 +1,9 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.autonomous;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
+
+import org.usfirst.frc.team166.robot.commands.wings.RaiseLeftWing;
+import org.usfirst.frc.team166.robot.commands.wings.RaiseRightWing;
 
 /**
  *
