+++ src/org/usfirst/frc/team166/robot/commands/autonomous/DriveDirection.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.autonomous;
 
 import edu.wpi.first.wpilibj.command.Command;
 
@@ -7,10 +7,13 @@ import org.usfirst.frc.team166.robot.Robot;
 /**
  *
  */
-public class CenterOnStep extends Command {
+public class DriveDirection extends Command {
+	private double angle;
 
-	public CenterOnStep() {
+	public DriveDirection(double driveAngle) {
+		// Use requires() here to declare subsystem dependencies
 		requires(Robot.drive);
+		driveAngle = angle;
 	}
 
 	// Called just before this Command runs the first time
@@ -21,20 +24,19 @@ public class CenterOnStep extends Command {
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
-		Robot.drive.centerOnStep();
+		Robot.drive.driveAngle(angle);
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		// return Robot.drive.isCentered();
 		return false;
 	}
 
 	// Called once after isFinished returns true
 	@Override
 	protected void end() {
-		Robot.drive.stopMotors();
+
 	}
 
 	// Called when another command which requires one or more of the same
