+++ src/org/usfirst/frc/team166/robot/commands/autonomous/DriveToStep.java
@@ -1,5 +1,6 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.autonomous;
 
+import edu.wpi.first.wpilibj.Preferences;
 import edu.wpi.first.wpilibj.command.Command;
 
 import org.usfirst.frc.team166.robot.Robot;
@@ -7,34 +8,36 @@ import org.usfirst.frc.team166.robot.Robot;
 /**
  *
  */
-public class OpenClaw extends Command {
+public class DriveToStep extends Command {
 
-	public OpenClaw() {
-		// Use requires() here to declare subsystem dependencies
-		// eg. requires(chassis);
-		requires(Robot.claw);
+	public DriveToStep() {
+		requires(Robot.drive);
 	}
 
 	// Called just before this Command runs the first time
 	@Override
 	protected void initialize() {
-		Robot.claw.open();
 	}
 
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
+		Robot.drive.printPDPAverageCurrent();
+		Robot.drive.driveForwardWithGyro();
+
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return true;
+		return (Robot.drive.getMotorCurrent() > Preferences.getInstance().getDouble("currentCutoff", 20));
 	}
 
 	// Called once after isFinished returns true
 	@Override
 	protected void end() {
+		Robot.drive.stopMotors();
+		Robot.drive.resetIntegral();
 	}
 
 	// Called when another command which requires one or more of the same
