+++ src/org/usfirst/frc/team166/robot/commands/drive/CancelDriveCommand.java
@@ -1,6 +1,5 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.drive;
 
-import edu.wpi.first.wpilibj.Preferences;
 import edu.wpi.first.wpilibj.command.Command;
 
 import org.usfirst.frc.team166.robot.Robot;
@@ -8,9 +7,10 @@ import org.usfirst.frc.team166.robot.Robot;
 /**
  *
  */
-public class DriveToStep extends Command {
+public class CancelDriveCommand extends Command {
 
-	public DriveToStep() {
+	public CancelDriveCommand() {
+		// Use requires() here to declare subsystem dependencies
 		requires(Robot.drive);
 	}
 
@@ -22,22 +22,17 @@ public class DriveToStep extends Command {
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
-		Robot.drive.printPDPAverageCurrent();
-		Robot.drive.driveForwardWithGyro();
-
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return (Robot.drive.getMotorCurrent() > Preferences.getInstance().getDouble("currentCutoff", 20));
+		return true;
 	}
 
 	// Called once after isFinished returns true
 	@Override
 	protected void end() {
-		Robot.drive.stopMotors();
-		Robot.drive.resetIntegral();
 	}
 
 	// Called when another command which requires one or more of the same
