+++ src/org/usfirst/frc/team166/robot/commands/drive/DriveWithJoysticks.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.drive;
 
 import edu.wpi.first.wpilibj.command.Command;
 
@@ -7,29 +7,32 @@ import org.usfirst.frc.team166.robot.Robot;
 /**
  *
  */
-public class LowerLeftWing extends Command {
+public class DriveWithJoysticks extends Command {
 
-	public LowerLeftWing() {
-		// Use requires() here to declare subsystem dependencies
-		// eg. requires(chassis);
-		requires(Robot.leftWing);
+	public DriveWithJoysticks() {
+		requires(Robot.drive);
 	}
 
 	// Called just before this Command runs the first time
 	@Override
 	protected void initialize() {
-		Robot.leftWing.lower();
 	}
 
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
+		Robot.drive.mecanumDrive(Robot.oi.getDriveJoystick());
+		Robot.drive.printEncoderValues();
+		Robot.drive.getGyro();
+		Robot.drive.getFrontDistance();
+		Robot.drive.getRightDistance();
+		Robot.drive.getLeftDistance();
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return true;
+		return false;
 	}
 
 	// Called once after isFinished returns true
