+++ src/org/usfirst/frc/team166/robot/commands/wings/LowerLeftWing.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.wings;
 
 import edu.wpi.first.wpilibj.command.Command;
 
@@ -7,36 +7,34 @@ import org.usfirst.frc.team166.robot.Robot;
 /**
  *
  */
-public class DriveDirection extends Command {
-	private double angle;
+public class LowerLeftWing extends Command {
 
-	public DriveDirection(double driveAngle) {
+	public LowerLeftWing() {
 		// Use requires() here to declare subsystem dependencies
-		requires(Robot.drive);
-		driveAngle = angle;
+		// eg. requires(chassis);
+		requires(Robot.leftWing);
 	}
 
 	// Called just before this Command runs the first time
 	@Override
 	protected void initialize() {
+		Robot.leftWing.lower();
 	}
 
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
-		Robot.drive.driveAngle(angle);
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return false;
+		return true;
 	}
 
 	// Called once after isFinished returns true
 	@Override
 	protected void end() {
-
 	}
 
 	// Called when another command which requires one or more of the same
