+++ src/org/usfirst/frc/team166/robot/commands/wings/LowerWings.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.wings;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
