+++ src/org/usfirst/frc/team166/robot/commands/wings/RaiseRightWing.java
@@ -1,39 +1,35 @@
-package org.usfirst.frc.team166.robot.commands;
+package org.usfirst.frc.team166.robot.commands.wings;
 
 import edu.wpi.first.wpilibj.command.Command;
 
 import org.usfirst.frc.team166.robot.Robot;
-import org.usfirst.frc.team166.robot.subsystems.Drive;
 
 /**
  *
  */
-public class StrafeDirection extends Command {
-	private Drive.StrafeDirection strafeDirection;
+public class RaiseRightWing extends Command {
 
-	public StrafeDirection(Drive.StrafeDirection direction) {
+	public RaiseRightWing() {
 		// Use requires() here to declare subsystem dependencies
 		// eg. requires(chassis);
-		requires(Robot.drive);
-		strafeDirection = direction;
+		requires(Robot.rightWing);
 	}
 
 	// Called just before this Command runs the first time
 	@Override
 	protected void initialize() {
+		Robot.rightWing.raise();
 	}
 
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
-		Robot.drive.strafeWithGyro(strafeDirection);
-		Robot.drive.printEncoderValues();
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return false;
+		return true;
 	}
 
 	// Called once after isFinished returns true
